refactor(credential): merge option_serde_secret into serde_secret::option submodule
Collapses two 20-line duplicate serde helper files into one. The non-option
variant stays at module root; the option variant moves into `pub mod option`
submodule. Attribute path `nebula_credential::option_serde_secret` ->
`nebula_credential::serde_secret::option`.
